attempt at fixing errors related to shape setup. Specifically on node.js where for example ImageObject is not avilable yet for Thing.image when the code was running. That should be fixed now.
If any more issues occur, we should: emit 1 global event when all shapes are ready. Then run all the decorators in the order we want (they are all registered somewhere) then continue with setting up the queries. Queries that run before the query is set up must wait for it to be ready